pyproject.toml
[project]
name = "multiagent-orchestrator"
version = "0.1.0"
description = "Plane-driven multi-agent orchestrator MVP"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn>=0.35.0",
  "httpx>=0.28.0",
  "pydantic>=2.11.0",
  "pydantic-settings>=2.10.0",
  "pyyaml>=6.0.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.0",
  "pytest-asyncio>=1.1.0",
  "pytest-cov>=6.2.0",
]

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["src"]
